% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_codes.R
\docType{data}
\name{lang_codes}
\alias{lang_codes}
\title{List of language codes}
\format{
An object of class \code{list} of length 133.
}
\usage{
lang_codes
}
\description{
This is the list of supported codes for languages included in gtranslate.
}
\references{
\url{https://cloud.google.com/translate/docs/languages}
}
\author{
Google
}
\keyword{datasets}
