% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_codes.R
\docType{data}
\name{lang_codes}
\alias{lang_codes}
\title{This list is the supported codes for language included in gtranslate}
\format{
An object of class \code{list} of length 133.
}
\usage{
lang_codes
}
\description{
This list is the supported codes for language included in gtranslate
}
\references{
\url{https://cloud.google.com/translate/docs/languages}
}
\author{
Google
}
\keyword{datasets}
